Moved Results logic to ResultsController and created ResultsService
