+++ src/DevAdventCalendarCompetition/DevAdventCalendarCompetition.Services/Interfaces/IResultsService.cs
@@ -1,0 +1,15 @@
+using System.Collections.Generic;
+using DevAdventCalendarCompetition.Repository.Models;
+using DevAdventCalendarCompetition.Services.Models;
+
+namespace DevAdventCalendarCompetition.Services.Interfaces
+{
+    public interface IResultsService
+    {
+        string PrepareUserEmailForRODO(string email);
+
+        UserPosition GetUserPosition(string userId);
+
+        Dictionary<int, List<TestResultDto>> GetAllTestResults();
+    }
+}
